Finally figured out what the 'stalls' were.
I feel stupid.

(cause: SDL batches up 'everything' without sending it to the OS
until we actually try to render it to the display, which meant
we could run through a ton of stuff without doing anything...
and then we'd flush a million texture updates at once. Not a good
way to try to deliver a semi-decent frame rate...)