Hack cursor restore
Without the ability to invert colours (SDL doesn't expose enough - need shader?),
I think the only right way to do this is to maintain a buffer of the actual
text (i.e. so we can redo the put, AND so that we can display the block
cursor properly with an invert, AND so we can cut/copy).

However, it's easy enough to just save the texture for now...